demo/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
opencv-python-headless
Pillow
mxnet
streamlit